Move helpdeskAccess into a custom useHelpdeskAccess hook, exported from the App component
